requirements.txt
streamlit==1.24.1
langchain==0.0.240
requests
python-dotenv
llamaapi
plotly
matplotlib
pandasai
judini==0.0.17